cephadm: move more funcs into data_utils.py
